Add precut flag to P700
With this change my P700 produces about 2cm of waste tape followed by
the desired label, with no additional padding.

As far as I know all of these printers have a cutter, so I am not sure
what determines which modules should have this flag set -- perhaps some
of them can cut at the end of a label only, not at the beginning?

We could perhaps have a commandline option to disable precut.